+++ src/modules/main_window.py
@@ -1,8 +1,6 @@
-import os
+import json
 import math
 import textwrap
-from time import sleep
-from PySide6 import QtCore
 from PySide6.QtWidgets import (
     QComboBox,
     QMainWindow,
@@ -22,16 +20,18 @@ from PySide6.QtGui import (
     QPainter,
     QPixmap,
 )
-from PySide6.QtCore import QCoreApplication, QEventLoop, QSize, Qt, QMarginsF, QRectF
+from PySide6.QtCore import QBuffer, QByteArray, QIODevice, QSize, Qt, QMarginsF, QRectF
 from pydantic import ValidationError
 
 from modules.config.config import Config
 from modules.editor_widget import EditorWidget
-from modules.persistence.file_manager import FileManager
+from modules.persistence.projects_service import ProjectsService
+from modules.persistence.file_service import FilesService
 from modules.shared.models.project_model import ProjectModel
 from modules.symbols.symbol_selector_widget import SymbolSelectorWindow
 from modules.symbols.symbol_mapper import SymbolMapper
 from modules.utils.logger import get_logger
+from modules.exceptions.exceptions import StorageError
 
 class MainWindow(QMainWindow):
     def __init__(self, config: Config, parent=None):
@@ -59,12 +59,13 @@ class MainWindow(QMainWindow):
         self.main_layout.setAlignment(self._editor, Qt.AlignmentFlag.AlignHCenter)
         self.main_layout.setContentsMargins(0, 20, 0, 20)
 
-        # Dependencia de persistencia
-        self.file_manager = FileManager()
-        self._editor.textChanged.connect(self.file_manager.set_to_unsaved)
+        # Dependencias de persistencia
+        self.projects_service = ProjectsService()
+        self._editor.textChanged.connect(self.projects_service.set_to_unsaved)
+        self.file_service = FilesService()
 
         # Prueba de ventana de drag y drop
-        self._symbol_collection_editor = SymbolSelectorWindow(self._symbol_mapper, self)
+        self._symbol_collection_editor = SymbolSelectorWindow(self._symbol_mapper, self.file_service, self)
         self._symbol_collection_editor.symbols_changed.connect(self.onSymbolsChanged)
 
         # comandos de usuario (QActions)
@@ -81,7 +82,6 @@ class MainWindow(QMainWindow):
         self.logger.info(f"Módulo Iniciado")           
 
     def onToggleViewChanged(self, checked: bool):  # CHECK
-        # print(self._toggleViewAction.isChecked())
         if checked:
             self._toggleViewAction.setText("Modo Texto")
             self._editor.switchToTextView()
@@ -89,7 +89,6 @@ class MainWindow(QMainWindow):
             self._toggleViewAction.setText("Modo Símbolos")
             self._editor.switchToImageView()
 
-    ## TODO: Agregar una función que rerenderice las imágenes, sin tener que pasar a texto y a imagen nuevamente
     def onFontSizeChanged(self):
 
         current_str = self._fontSizeBox.currentText()
@@ -104,109 +103,131 @@ class MainWindow(QMainWindow):
                 self._fontSizeBox.setCurrentText(new_size_str)
             new_size = int(
                 float(new_size_str)
-            )  # Falta manejar excepciones o casos en que sea un str que no pueda convertirse a int
+            )
             self._editor.setFontSize(new_size)
 
             self.updateSymbolsView()
         self._editor.setFocus()
 
+
+    ## TODO: Una función que renderice las imágenes sin tener que pasar todo a modo texto antes
     def updateSymbolsView(self):
+        """Actualiza el dibujo de los símbolos en el editor"""
         if not self._toggleViewAction.isChecked():
             self._editor.switchToTextView()
             self._editor.switchToImageView()        
 
     def onSymbolsChanged(self):
         self.updateSymbolsView()
-        self.file_manager.set_to_unsaved()
+        self.projects_service.set_to_unsaved()
 
     def onOpenFile(self):  ## CHECK (Solo falta revisar el paso de los switchs)
-
+        """Abre un archivo de proyecto en la ruta señalada por el usuario y:
+            - Carga su contenido en el editor       
+            - Carga la colección de símbolos del archivo (si no la encuentra lo anuncia)        
+            - Ajusta el tamaño de la fuente según el indicado en el archivo
+        """
         file_name, _ = QFileDialog.getOpenFileName(
             self, "Abrir Proyecto", "", "Archivo de Proyecto (*.json)"
         )
         if not file_name:
             return None
         try:
-            project = self.file_manager.openFile(file_name)
-            collection = self._symbol_collection_editor.select_collection_by_name(
-                project.collectionName
-            )
-            if project.imageSize is not None:
+            project = self.projects_service.open(file_name)
+            
+            if project.imageSize:
                 self._fontSizeBox.setCurrentText(f"{project.imageSize}")
                 self._editor.setFontSize(project.imageSize)
-
-            if not collection:
-                self.logger.warning(f"No se encontró la colección '{project.collectionName}'")
-                QMessageBox.information(self, "Proyecto sin símbolos", f"No se encontró la colección de símbolos '{project.collectionName}' del proyecto. Puedes elegir una o crear una nueva en 'Elegir Símbolos'")
-
-            if self._toggleViewAction.isChecked():
-                self._editor.setContent(project.content)
+            
+            if project.collectionName:
+                collection = self._symbol_collection_editor.select_collection_by_name(
+                    project.collectionName
+                )
+                if not collection:
+                    self.logger.warning(f"No se encontró la colección '{project.collectionName}'")
+                    QMessageBox.information(self, "Colección de simbolos no encontrada", f"No se encontró la colección de símbolos '{project.collectionName}' del proyecto. Puedes elegir una o crear una nueva en 'Elegir Símbolos'")
             else:
-                self._editor.switchToTextView()
-                self._editor.setContent(project.content)
-                self._editor.switchToImageView()
+                QMessageBox.information(self, "Proyecto sin símbolos", f"El proyecto no tiene una colección de símbolos definida. Puedes elegir una o crear una nueva en 'Elegir Símbolos'")
+            self._editor.setContent(project.content)        
             self.logger.info(f"Proyecto abierto: '{file_name}'")
-        except ValueError as e:
-            self.logger.error(f"No se pudo abrir el archivo: {str(e)}")
-            QMessageBox.critical(self, "Error", f"El archivo no es válido o está corrupto: {file_name}")
 
-        except ValidationError as e:
-            self.logger.error(f"El archivo no es válido o está corrupto: {file_name}")
+        except (ValidationError, json.JSONDecodeError) as e:
+            self.logger.warning(f"El archivo no es válido o está corrupto: {file_name} | {e}")
             QMessageBox.critical(
-                self, "Error", f"El archivo no es válido o está corrupto"
+                self, "Error de archivo", f"El archivo no es válido o está corrupto."
             )
-
+        except FileNotFoundError as e:
+            self.logger.warning(f"El archivo no es válido o está corrupto: {file_name} | {e}")
+            QMessageBox.critical(
+                self, "Error de archivo", f"Archivo no encontrado: {file_name}."
+            )
+        except StorageError as e:
+            self.logger.error(f"No se pudo abrir el archivo {file_name} | {e}")
+            QMessageBox.critical(
+                self, "Error de acceso", f"No se pudo abrir el archivo: {file_name}. Verifica los permisos del sistema."
+            )
         except Exception as e:
-            self.logger.error(f"No se pudo abrir el archivo: {file_name}")
+            self.logger.critical(f"Excepción no controlada al abrir archivo: {file_name} | {e}")
             QMessageBox.critical(
-                self, "Error", f"No se pudo abrir el archivo: {str(e)}"
-            )
-        self.file_manager.set_to_saved()
+                self, "Error misterioso", "Ocurrió un error inesperado en la aplicación."
+            )            
+        self.projects_service.set_to_saved()
 
     def onSaveFile(
         self,
-    ):  # CHECK. Falta pasar la lógica de obtención del contenido a la clase del editor
-        current_filename = self.file_manager.get_current_filename()
-        if current_filename is None:
+    ):
+        current_filepath = self.projects_service.get_current_filepath()
+        if current_filepath is None:
             self.onSaveFileAs()
             return
-        project = self._projectModel()
+        project = self._generateProjectModel()
         try:
-            self.file_manager.saveFile(project)
-            self.logger.info(f"Proyecto guardado: '{current_filename}'")
-        except Exception as error:
+            self.projects_service.save(project)
+            self.logger.info(f"Proyecto guardado: '{current_filepath}'")
+            self.projects_service.set_to_saved()
+        except StorageError as e:
+            self.logger.error(f"No se pudo guardar el archivo {current_filepath} | {e}")            
             QMessageBox.critical(
-                self, "Error", f"No se pudo guardar el archivo: {str(error)}"
+                self, "Error al guardar", f"No se pudo guardar el archivo: {current_filepath}. \nVerifica que tengas espacio disponible en disco y permisos para guardar."
             )
-        self.file_manager.set_to_saved()
+        except Exception as e:
+            self.logger.critical(f"Error inesperado al guardar el archivo: {current_filepath} | {e}")
+            QMessageBox.critical(
+                self, "Error misterioso", "Ocurrió un error inesperado en la aplicación al intentar guardar el archivo."
+            ) 
     def onSaveFileAs(
         self,
-    ):  # CHECK. Falta pasar la lógica de obtención del contenido a la clase del editor
+    ):
 
-        project = self._projectModel()
-        current_filename = self.file_manager.get_current_filename()
-        new_filename = f"{current_filename if current_filename is not None else self.config.UNTITLED_DEFAULT_FILENAME}{self.file_manager.get_file_extension()}"
+        project = self._generateProjectModel()
+        current_filename = self.projects_service.get_current_filename()
+        new_filename = f"{current_filename if current_filename is not None else self.config.UNTITLED_DEFAULT_FILENAME}{self.projects_service.get_file_extension()}"
         file_name, _ = QFileDialog.getSaveFileName(
             self,
             "Guardar Proyecto",
             new_filename,
-            f"Archivo de Proyecto (*{self.file_manager.get_file_extension()})",
+            f"Archivo de Proyecto (*{self.projects_service.get_file_extension()})",
         )
         if not file_name:
             return
-
         try:
-            self.file_manager.saveFileAs(file_name, project)
+            self.projects_service.save(project, file_name)
             self.logger.info(f"Proyecto guardado como: '{file_name}'")
-        except Exception as error:
+            self.projects_service.set_to_saved()
+        except StorageError as e:
+            self.logger.error(f"No se pudo guardar el archivo {file_name} | {e}")            
             QMessageBox.critical(
-                self, "Error", f"No se pudo crear el archivo de proyecto: {str(error)}"
+                self, "Error al guardar", f"No se pudo guardar el archivo: {file_name}. \nVerifica que tengas espacio disponible en disco y permisos para guardar."
             )
-        self.file_manager.set_to_saved()
+        except Exception as e:
+            self.logger.critical(f"Error inesperado al guardar el archivo: {file_name} | {e}")
+            QMessageBox.critical(
+                self, "Error misterioso", "Ocurrió un error inesperado en la aplicación al intentar guardar el archivo."
+            ) 
 
     def onExportPdf(self):
 
-        current_filename = self.file_manager.get_current_filename()
+        current_filename = self.projects_service.get_current_filename()
         file_name, _ = QFileDialog.getSaveFileName(
             self,
             "Exportar a PDF",
@@ -219,21 +240,25 @@ class MainWindow(QMainWindow):
         if not file_name.lower().endswith(".pdf"):
             file_name += ".pdf"
 
-        if self._toggleViewAction.isChecked():
-            self._toggleViewAction.setChecked(False)
-
-        self._editor.prepare_for_export(factor=4)
+        # if self._toggleViewAction.isChecked():
+        #     self._toggleViewAction.setChecked(False)
+        if not self._editor._is_text_view_mode:
+            self._editor.prepare_for_export(factor=4)
 
         page_height = self._editor.page_height
         page_width = self._editor.page_width
 
-        pdf_writer = QPdfWriter(file_name)
-        pdf_writer.setPageSize(QSize(page_width, page_height))
-        pdf_writer.setPageMargins(QMarginsF(0, 0, 0, 0))
-        pdf_writer.setResolution(72)  # 72 DPI es el estándar nativo de puntos para PDF
+        bytes_array = QByteArray()
+        buffer = QBuffer(bytes_array)
+        buffer.open(QIODevice.OpenModeFlag.WriteOnly)
+        pdf_bytes_writer = QPdfWriter(buffer)
+        pdf_bytes_writer.setPageSize(QSize(page_width, page_height))
+        pdf_bytes_writer.setPageMargins(QMarginsF(0, 0, 0, 0))
+        pdf_bytes_writer.setResolution(72)  # 72 DPI es el estándar nativo de puntos para PDF
+        pdf_bytes_writer.setCreator(self.config.APP_NAME)
         
         painter = QPainter()
-        if not painter.begin(pdf_writer):
+        if not painter.begin(pdf_bytes_writer):
             QMessageBox.critical(self, "Error", "No se pudo activar el PDF.")
             self._editor.restore_after_export()
             return
@@ -246,52 +271,12 @@ class MainWindow(QMainWindow):
 
         for i in range(total_pages):
             if i > 0:
-                pdf_writer.newPage()
+                pdf_bytes_writer.newPage()
 
             painter.save()
             painter.translate(0, -(i * page_height))
             painter.setClipRect(0, i * page_height, page_width, page_height)
-
-            self._editor.document().drawContents(
-                painter, QRectF(0, 0, page_width, total_height)
-            )
-            painter.restore()
-
-        painter.end()
-        self._editor.restore_after_export()
-
-    def onExportImage(self):
-        file_name, _ = QFileDialog.getSaveFileName(
-            self,
-            "Exportar como Imagen",
-            f"{self.file_manager.get_current_filename()}.png",
-            "Imagen PNG (*.png)",
-        )
-        if not file_name:
-            return
-
-        base_path, _ = os.path.splitext(file_name)
-
-        image_mode = not self._toggleViewAction.isChecked()
-        sufix = "imagen" if image_mode else "texto"
-
-        if image_mode:
-            self._editor.prepare_for_export(factor=4)
-
-        page_height = self._editor.page_height
-        page_width = self._editor.page_width
-        total_height = self._editor.document().size().height()
-        total_pages = max(1, math.ceil(float(total_height) / page_height))
-        for i in range(total_pages):
-            pixmap = QPixmap(page_width, page_height)
-            # pixmap.fill(Qt.GlobalColor.white)
-            pixmap.fill(Qt.GlobalColor.transparent)
-
-            painter = QPainter(pixmap)
-            painter.save()
-            painter.translate(0, -(i * page_height))
-            painter.setClipRect(0, i * page_height, page_width, page_height)
-            if image_mode:
+            if not self._editor._is_text_view_mode:
                 self._editor.document().drawContents(
                     painter, QRectF(0, 0, page_width, total_height)
                 )
@@ -301,21 +286,158 @@ class MainWindow(QMainWindow):
                 ctx.clip = QRectF(0, 0, page_width, total_height)
                 self._editor.document().documentLayout().draw(painter, ctx)
             painter.restore()
-            painter.end()
 
-            page_file_name = f"{base_path}_{i + 1}_{sufix}.png"
-            pixmap.save(page_file_name, "PNG")
+        painter.end()
+        buffer.close()
+        try:
+            self.file_service.save(file_name, bytes_array)
+        except StorageError as e:
+            self.logger.error(f"No se pudo exportar el PDF {file_name} | {e}")            
+            QMessageBox.critical(
+                self, "Error al exportar", f"No se pudo exportar el PDF: {file_name}. \nVerifica que tengas espacio disponible en disco y permisos para guardar."
+            )
+        except Exception as e:
+            self.logger.critical(f"Error inesperado al exportar el PDF: {file_name} | {e}")
+            QMessageBox.critical(
+                self, "Error misterioso", "Ocurrió un error inesperado en la aplicación al intentar exportar el PDF."
+            ) 
+        finally:
+            if not self._editor._is_text_view_mode:            
+                self._editor.restore_after_export()
 
-        if image_mode:
-            self._editor.prepare_for_export(factor=4)
+    def onExportImage(self):
+        file_name, _ = QFileDialog.getSaveFileName(
+            self,
+            "Exportar como Imagen",
+            f"{self.projects_service.get_current_filename()}.png",
+            "Imagen PNG (*.png)",
+        )
+        if not file_name:
+            return
+        definition_multiplier = 6
+        if not self._editor._is_text_view_mode:
+            self._editor.prepare_for_export(factor=definition_multiplier)
+        byte_array = QByteArray()
+        buffer = QBuffer(byte_array)
+        buffer.open(QIODevice.OpenModeFlag.WriteOnly)
+
+        page_height = self._editor.page_height
+        page_width = self._editor.page_width
+        total_height = self._editor.document().size().height()
+
+
+        ## Modo Horizontal
+        ## Ordenar las páginas en un lienzo horizontal
+        total_pages = max(1, math.ceil(float(total_height) / page_height))
+        pixmap = QPixmap(page_width*definition_multiplier*total_pages, page_height*definition_multiplier)
+        pixmap.fill(Qt.GlobalColor.transparent)
+        painter = QPainter(pixmap)
+        painter.scale(definition_multiplier, definition_multiplier)
+        for i in range(total_pages):
+            painter.save()
+            painter.translate(i*page_width, -(i*page_height))
+            clip_rect = QRectF(0, page_height * i, page_width, page_height)
+            painter.setClipRect(clip_rect)            
+            if not self._editor._is_text_view_mode:
+                self._editor.document().drawContents(
+                    painter, clip_rect
+                )
+            else:
+                ctx = QAbstractTextDocumentLayout.PaintContext()
+                ctx.palette.setColor(QPalette.ColorGroup.All, QPalette.ColorRole.Text, "#333333")
+                ctx.clip =clip_rect
+                self._editor.document().documentLayout().draw(painter, ctx)
+            painter.restore()
+        painter.end()
+        pixmap.save(buffer, "PNG") 
+        buffer.close()
+        try:
+            self.file_service.save(file_name, byte_array)
+        except StorageError as e:
+            self.logger.error(f"No se pudo exportar el PDF {file_name} | {e}")            
+            QMessageBox.critical(
+                self, "Error al exportar", f"No se pudo exportar el PDF: {file_name}. \nVerifica que tengas espacio disponible en disco y permisos para guardar."
+            )
+        except Exception as e:
+            self.logger.critical(f"Error inesperado al exportar el PDF: {file_name} | {e}")
+            QMessageBox.critical(
+                self, "Error misterioso", "Ocurrió un error inesperado en la aplicación al intentar exportar el PDF."
+            ) 
+        finally:
+            if not self._editor._is_text_view_mode:            
+                self._editor.restore_after_export()
+        ## Modo Vertical
+        ## Ordenar las páginas en un lienzo Vertical
+        # pixmap = QPixmap(page_width*definition_multiplier, total_height*definition_multiplier)
+        # pixmap.fill(Qt.GlobalColor.transparent)
+        # painter = QPainter(pixmap)
+        # painter.scale(definition_multiplier, definition_multiplier)
+        # painter.save()
+        # # painter.translate(0, 0)
+        # painter.setClipRect(0, 0, page_width*definition_multiplier, total_height*definition_multiplier)
+        # if not self._editor._is_text_view_mode:
+        #     self._editor.document().drawContents(
+        #         painter, QRectF(0, 0, page_width, total_height)
+        #     )
+        # else:
+        #     ctx = QAbstractTextDocumentLayout.PaintContext()
+        #     ctx.palette.setColor(QPalette.ColorGroup.All, QPalette.ColorRole.Text, "#333333")
+        #     ctx.clip = QRectF(0, 0, page_width, total_height)
+        #     self._editor.document().documentLayout().draw(painter, ctx)
+        # painter.restore()
+        # painter.end()
+        # buffer.close()
+        # pixmap.save(buffer, "PNG") # Retorna un bool
+        # try:
+        #   self.file_service.save(file_name, byte_array)
+        # except StorageError as e:
+        #     self.logger.error(f"No se pudo exportar el PDF {file_name} | {e}")            
+        #     QMessageBox.critical(
+        #         self, "Error al exportar", f"No se pudo exportar el PDF: {file_name}. \nVerifica que tengas espacio disponible en disco y permisos para guardar."
+        #     )
+        # except Exception as e:
+        #     self.logger.critical(f"Error inesperado al exportar el PDF: {file_name} | {e}")
+        #     QMessageBox.critical(
+        #         self, "Error misterioso", "Ocurrió un error inesperado en la aplicación al intentar exportar el PDF."
+        #     ) 
+        # finally:
+        #     if not self._editor._is_text_view_mode:            
+        #         self._editor.restore_after_export()
+
+        ## Modo Múltiples archivos
+        ## Ordenar las páginas en un lienzo por página
+        # total_pages = max(1, math.ceil(float(total_height) / page_height))
+        # for i in range(total_pages):
+        #     pixmap = QPixmap(page_width, page_height)
+        #     # pixmap.fill(Qt.GlobalColor.white)
+        #     pixmap.fill(Qt.GlobalColor.transparent)
+
+        #     painter = QPainter(pixmap)
+        #     painter.save()
+        #     painter.translate(0, -(i * page_height))
+        #     painter.setClipRect(0, i * page_height, page_width, page_height)
+        #     if not self._editor._is_text_view_mode:
+        #         self._editor.document().drawContents(
+        #             painter, QRectF(0, 0, page_width, total_height)
+        #         )
+        #     else:
+        #         ctx = QAbstractTextDocumentLayout.PaintContext()
+        #         ctx.palette.setColor(QPalette.ColorGroup.All, QPalette.ColorRole.Text, "#333333")
+        #         ctx.clip = QRectF(0, 0, page_width, total_height)
+        #         self._editor.document().documentLayout().draw(painter, ctx)
+        #     painter.restore()
+        #     painter.end()
+
+        #     page_file_name = f"{base_path}_{i + 1}_{sufix}.png"
+        #     pixmap.save(page_file_name, "PNG")
+
 
     def _open_symbols_window(self):
+        """Muestra la ventana del editor de colecciones de símbolos"""
         self._symbol_collection_editor.show()
 
-
-
     def closeEvent(self, event):
-        if not self.file_manager.is_saved():
+        if not self.projects_service.is_saved():
             button = QMessageBox.question(
                 self,
                 "Cambios no guardados",
@@ -337,6 +459,7 @@ class MainWindow(QMainWindow):
         return super().closeEvent(event)
 
     def _about_window(self):
+        """Despliega una ventana informativa con datos de la"""
 
         QMessageBox().information(
             self,
@@ -351,7 +474,7 @@ class MainWindow(QMainWindow):
         )
 
     def _tutorial_window(self):
-
+        """Despliega una ventana informativa con instrucciones de uso"""
         QMessageBox().information(
             self,
             "Instrucciones",
@@ -363,7 +486,7 @@ class MainWindow(QMainWindow):
             <b>Cambiar modo de vista</b><br>
 
             Presiona el botón 'Modo Texto' o 'Modo Símbolos' para cambiar el modo en que se muestran los símbolos en la hoja.<br>
-            """),
+            """)
         )
     
     def _create_actions(self):
@@ -473,17 +596,11 @@ class MainWindow(QMainWindow):
         box.currentIndexChanged.connect(self.onFontSizeChanged)
         return box
 
-    def _projectModel(self) -> ProjectModel:  # Auxiliar, crea el objeto de guardado
-
-        was_in_images_mode = not self._toggleViewAction.isChecked()
-        if was_in_images_mode:
-            self._editor.switchToTextView()
+    def _generateProjectModel(self) -> ProjectModel:  # Auxiliar, crea el objeto de guardado
         project = ProjectModel(
             version=self.config.APP_VERSION,
-            content=self._editor.toPlainText(),
+            content=self._editor.getContent(),
             imageSize=int(self._fontSizeBox.currentText()),
             collectionName=self._symbol_collection_editor.get_current_collection_name(),
         )
-        if was_in_images_mode:
-            self._editor.switchToImageView()
         return project