+++ src/modules/shared/models/project_model.py
@@ -5,5 +5,5 @@ class ProjectModel(BaseModel):
 
     version: str
     content: str
-    imageSize: int
-    collectionName: str
+    imageSize: int | None = None
+    collectionName: str = ""