+++ src/modules/symbols/symbol_drop_surface.py
@@ -42,22 +42,29 @@ class SymbolDropSurface(QWidget):
             self.image_dropped.emit()
             event.accept()
 
+    def get_file_name(self):
+        return f"{self.symbol_name}{Path(self.image_path).suffix.lower()}"
+
+    def get_format(self):
+        """Retorna el formato de la imagen como str"""
+        return Path(self.image_path).suffix.replace(".", "").upper()
+
     def clear(self):
         """Asigna a la superficie un Pixmap vacío y la actualiza."""
         self.pixmap = QPixmap()
         self.update()
 
-    def save_to_file(self, symbol_dir_path: Path) -> bool:
-        """Guarda la imagen con el nombre del símbolo y la extensión de la imagen de origen en el directorio dado."""
-        if self.pixmap and self.has_symbol():
-            filename = f"{self.symbol_name}{Path(self.image_path).suffix.lower()}"
-            image_path = symbol_dir_path.joinpath(filename)
-            saved = self.pixmap.save(str(image_path))
-            if saved:
-                self.image_path = image_path
-            return saved
-        else:
-            return False
+    # def save_to_file(self, symbol_dir_path: Path) -> bool:
+    #     """Guarda la imagen con el nombre del símbolo y la extensión de la imagen de origen en el directorio dado."""
+    #     if self.pixmap and self.has_symbol():
+    #         filename = f"{self.symbol_name}{Path(self.image_path).suffix.lower()}"
+    #         image_path = symbol_dir_path.joinpath(filename)
+    #         saved = self.pixmap.save(str(image_path))
+    #         if saved:
+    #             self.image_path = image_path
+    #         return saved
+    #     else:
+    #         return False
 
     def set_symbol(self, image_path: str, crop: bool = True):
         """Instancia un QPixmap a partir de una ruta de image y la fija en el Drop Surface.